v_0_0_29: update text type to double for integer

+++ app/src/main/java/com/example/sabencostimes/domain/NYTMarketApiDomain.kt
@@ -1,3 +1,3 @@
 package com.example.sabencostimes.domain
 
-data class NYTMarketApiDomain(val name:String,val points:Long,val percentageChange:Long, val timestamp:String)
+data class NYTMarketApiDomain(val name:String,val points:Double,val percentageChange:Double, val timestamp:String)
